Look for presence of disease

Look for the presence of disease over a period that follows a given time point.

any_disease(x, n = 1)

Arguments

x
A boolean vector informing about the presence or absence of disease at a given point in time. This implies that this vector is considered to be chronologically ordered, the most recent observations being first.

n
A integer scalar telling the number of time points after the current one to look for for defining the period of time over which we want to look for presence of disease.

Value

A boolean of vector of the same length as x, the n final values of which are NA.

Examples

disease <- rep(rep(c(FALSE, TRUE), 7),
               c(rep(1, 4), 2, 3, 2, 2, 1, 1, 3, 2, 2, 3))
cbind(disease, any_disease(disease, 3))

#>       disease      
#>  [1,]   FALSE  TRUE
#>  [2,]    TRUE  TRUE
#>  [3,]   FALSE  TRUE
#>  [4,]    TRUE  TRUE
#>  [5,]   FALSE  TRUE
#>  [6,]   FALSE  TRUE
#>  [7,]    TRUE  TRUE
#>  [8,]    TRUE  TRUE
#>  [9,]    TRUE  TRUE
#> [10,]   FALSE  TRUE
#> [11,]   FALSE  TRUE
#> [12,]    TRUE  TRUE
#> [13,]    TRUE  TRUE
#> [14,]   FALSE  TRUE
#> [15,]    TRUE FALSE
#> [16,]   FALSE  TRUE
#> [17,]   FALSE  TRUE
#> [18,]   FALSE  TRUE
#> [19,]    TRUE  TRUE
#> [20,]    TRUE  TRUE
#> [21,]   FALSE  TRUE
#> [22,]   FALSE  TRUE
#> [23,]    TRUE    NA
#> [24,]    TRUE    NA
#> [25,]    TRUE    NA