Fill gaps

This function fills the gaps in a sequence of booleans.

fill_gaps(x, n = 1)

Arguments

x
A vector of boolean.

n
A scalar integer specifying the length of gaps that we want to fill.

Value

A vector of boolean of the same size as x.

Details

If n = 1, it will replace any TRUE FALSE TRUE sequence by TRUE TRUE TRUE. If n = 1, it will additionally replace any TRUE FALSE FALSE TRUE sequence by TRUE TRUE TRUE TRUE and so on.

Examples

x <- rep(c(rep(c(FALSE, TRUE), 7), FALSE),
         c(1, 3, 1, 1, 1, 2, 1, 2, 3, 2, 3, 1, 2, 3, 1))
fill_gaps(x)

#> Le chargement a nécessité le package : magrittr

#>  [1] FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [13] FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE
#> [25]  TRUE  TRUE FALSE

fill_gaps(x, 2)

#>  [1] FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [13] FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE
#> [25]  TRUE  TRUE FALSE